+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -10,6 +10,7 @@ import com.nhaarman.mockitokotlin2.inOrder
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.mockingDetails
 import com.nhaarman.mockitokotlin2.never
+import com.nhaarman.mockitokotlin2.spy
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
@@ -1038,39 +1039,38 @@ class SentryClientTest {
         fixture.getSut().captureTransaction(transaction, createScopeWithAttachments(), null)
 
         verifyAttachmentsInEnvelope(transaction.eventId)
-        assertFails { verifyProfilingTraceInEnvelope(transaction.eventId) }
+        assertFails { verifyProfilingTraceInEnvelope(SentryId(fixture.profilingTraceData.profileId)) }
     }
 
     @Test
-    fun `when captureTransaction with attachments and profiling data`() {
-        val transaction = SentryTransaction(fixture.sentryTracer)
-        fixture.getSut().captureTransaction(transaction, null, createScopeWithAttachments(), null, fixture.profilingTraceData)
-        verifyAttachmentsInEnvelope(transaction.eventId)
-        verifyProfilingTraceInEnvelope(transaction.eventId)
+    fun `when captureProfile`() {
+        val client = fixture.getSut()
+        client.captureProfile(fixture.profilingTraceData)
+        verifyProfilingTraceInEnvelope(SentryId(fixture.profilingTraceData.profileId))
     }
 
     @Test
-    fun `when captureTransaction with profiling data`() {
+    fun `captureTransaction with profile calls captureProfile and captureTransaction`() {
         val transaction = SentryTransaction(fixture.sentryTracer)
-        fixture.getSut().captureTransaction(transaction, null, null, null, fixture.profilingTraceData)
-        assertFails { verifyAttachmentsInEnvelope(transaction.eventId) }
-        verifyProfilingTraceInEnvelope(transaction.eventId)
+        val client = spy(fixture.getSut())
+        client.captureTransaction(transaction, null, null, null, fixture.profilingTraceData)
+        verify(client).captureTransaction(transaction, null, null, null)
+        verify(client).captureProfile(fixture.profilingTraceData)
     }
 
     @Test
-    fun `when captureTransaction with empty trace file, profiling data is not sent`() {
-        val transaction = SentryTransaction(fixture.sentryTracer)
-        fixture.getSut().captureTransaction(transaction, null, null, null, fixture.profilingTraceData)
+    fun `when captureProfile with empty trace file, profile is not sent`() {
+        val client = fixture.getSut()
+        client.captureProfile(fixture.profilingTraceData)
         fixture.profilingTraceFile.writeText("")
-        assertFails { verifyProfilingTraceInEnvelope(transaction.eventId) }
+        assertFails { verifyProfilingTraceInEnvelope(SentryId(fixture.profilingTraceData.profileId)) }
     }
 
     @Test
-    fun `when captureTransaction with non existing profiling trace file, profiling trace data is not sent`() {
-        val transaction = SentryTransaction(fixture.sentryTracer)
-        fixture.getSut().captureTransaction(transaction, null, createScopeWithAttachments(), null, fixture.profilingNonExistingTraceData)
-        verifyAttachmentsInEnvelope(transaction.eventId)
-        assertFails { verifyProfilingTraceInEnvelope(transaction.eventId) }
+    fun `when captureProfile with non existing profiling trace file, profile is not sent`() {
+        val client = fixture.getSut()
+        client.captureProfile(fixture.profilingNonExistingTraceData)
+        assertFails { verifyProfilingTraceInEnvelope(SentryId(fixture.profilingTraceData.profileId)) }
     }
 
     @Test
@@ -2046,8 +2046,7 @@ class SentryClientTest {
                     item.header.type == SentryItemType.Profile
                 }
                 assertNotNull(profilingTraceItem?.data)
-            },
-            anyOrNull()
+            }
         )
     }